+++ config/splat.pspeu.strno3.yaml
@@ -67,7 +67,7 @@ segments:
       - [0x19FB8, c, rno3/prim_helpers]
       - [0x1A748, c, rno3_psp/unk_1A748]
       - [0x1C7D8, c, rno3/e_jewel_sword_puzzle]
-      - [0x1D498, c, rno3_psp/e_stairway]
+      - [0x1D498, c, rno3/e_stairway]
       - [0x1DD38, c, rno3/e_background_bushes_trees]
       - [0x1E640, c, rno3_psp/st_init]
       - [0x1E9A8, c, rno3/e_breakable]
