+++ src/st/rno3/e_stairway.c
@@ -1,6 +1,202 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "rno3.h"
 
-INCLUDE_ASM("st/rno3/nonmatchings/e_stairway", EntityStairwayPiece);
+#ifdef VERSION_PSP
+extern s32 E_ID(FALLING_ROCK);
+#endif
 
-INCLUDE_ASM("st/rno3/nonmatchings/e_stairway", EntityFallingRock);
+extern EInit D_us_80180A34;
+extern EInit g_EInitInteractable;
+
+// Stairway piece you can break before Death encounter
+void EntityStairwayPiece(Entity* self) {
+    u16 page;
+    u16 clut;
+    u8* pagePtr;
+    u8* indexPtr;
+    u8* clutPtr;
+
+    u8 left;
+    u8 top;
+    u16 packed_uv;
+
+    Primitive* prim;
+    Entity* newEntity;
+    Collider collider;
+    s32 primIndex;
+    s16 x, y;
+    s32 i;
+
+    switch (self->step) {
+    case 0:
+        InitializeEntity(g_EInitInteractable);
+        self->hitboxWidth = 8;
+        self->hitboxHeight = 8;
+        self->posX.i.hi = 104 - g_Tilemap.scrollX.i.hi;
+        self->posY.i.hi = 56 - g_Tilemap.scrollY.i.hi;
+        self->hitPoints = 16;
+        if (g_CastleFlags[INV_DEATH_STAIRWAY_BROKEN]) {
+            self->hitboxState = 0;
+            g_Tilemap.fg[0x126] = 0x3EE;
+            g_Tilemap.fg[0xC6] = 0x3D2;
+            self->step = 32;
+            break;
+        }
+        self->hitboxState = 2;
+        g_Tilemap.fg[0x126] = 0x408;
+        g_Tilemap.fg[0xC6] = 0x40D;
+        break;
+
+    case 1:
+        if (self->hitFlags) {
+            g_api.PlaySfx(SFX_STOMP_SOFT_A);
+        }
+
+        if (self->flags & FLAG_DEAD) {
+            self->step++;
+        }
+        break;
+
+    case 2:
+        g_api.PlaySfx(SFX_WALL_DEBRIS_B);
+        g_Tilemap.fg[0x126] = 0x3EE;
+        g_Tilemap.fg[0xC6] = 0x3D2;
+        g_CastleFlags[INV_DEATH_STAIRWAY_BROKEN] = true;
+
+        newEntity = AllocEntity(&g_Entities[160], &g_Entities[192]);
+        if (newEntity != NULL) {
+            CreateEntityFromEntity(E_HEART_DROP, self, newEntity);
+            newEntity->params = self->params;
+        }
+
+        newEntity = AllocEntity(&g_Entities[224], &g_Entities[256]);
+        if (newEntity != NULL) {
+            CreateEntityFromEntity(E_INTENSE_EXPLOSION, self, newEntity);
+            newEntity->params = 0x10;
+            newEntity->zPriority = self->zPriority + 1;
+            newEntity->posX.i.hi += 8;
+            newEntity->posY.i.hi += 8;
+        }
+
+        primIndex = g_api.AllocPrimitives(PRIM_GT4, 16);
+        if (primIndex == -1) {
+            DestroyEntity(self);
+            return;
+        }
+        self->flags |= FLAG_HAS_PRIMS;
+        self->primIndex = primIndex;
+        prim = &g_PrimBuf[primIndex];
+        self->ext.prim = prim;
+        UnkPolyFunc2(prim);
+        pagePtr = g_Tilemap.tileDef->gfxPage;
+        indexPtr = g_Tilemap.tileDef->gfxIndex;
+        clutPtr = g_Tilemap.tileDef->clut;
+
+        packed_uv = indexPtr[0x409];
+        left = packed_uv << 4;
+        top = packed_uv & 0xF0;
+        page = pagePtr[0x409];
+        clut = clutPtr[0x409];
+        prim->clut = clut;
+        prim->tpage = page + 8;
+
+        prim->u0 = prim->u2 = left;
+        prim->u1 = prim->u3 = left + 15;
+        prim->v0 = prim->v1 = top;
+        prim->v2 = prim->v3 = top + 15;
+        prim->next->x1 = self->posX.i.hi;
+        prim->next->y0 = self->posY.i.hi;
+        LOW(prim->next->u0) = 0x10000;
+        LOW(prim->next->r1) = 0xFFFF0000;
+        LOH(prim->next->r2) = 16;
+        LOH(prim->next->b2) = 16;
+        prim->priority = self->zPriority;
+        prim->drawMode = DRAW_UNK02;
+        prim = prim->next;
+        self->step++;
+
+    case 3:
+        prim = self->ext.prim;
+        LOH(prim->next->tpage) -= 0x20; // has to be loaded signed on psp
+        LOW(prim->next->r1) += 0x2000;
+        UnkPrimHelper(prim);
+        x = prim->next->x1;
+        y = prim->next->y0 + 8;
+        g_api.CheckCollision(x, y, &collider, 0);
+        if (collider.effects & EFFECT_SOLID) {
+            self->posX.i.hi = x;
+            self->posY.i.hi = y - 12;
+            self->step++;
+        }
+        break;
+
+    case 4:
+        g_api.PlaySfx(SFX_WALL_DEBRIS_B);
+        newEntity = AllocEntity(&g_Entities[224], &g_Entities[256]);
+        if (newEntity != NULL) {
+            CreateEntityFromEntity(E_EXPLOSION, self, newEntity);
+            newEntity->params = 0x11;
+            newEntity->zPriority = self->zPriority + 1;
+        }
+
+        for (i = 0; i < 6; i++) {
+            newEntity = AllocEntity(&g_Entities[224], &g_Entities[256]);
+            if (newEntity != NULL) {
+                CreateEntityFromEntity(E_ID(FALLING_ROCK), self, newEntity);
+                newEntity->params = Random() & 3;
+                if (newEntity->params == 3) {
+                    newEntity->params = 0;
+                }
+            }
+        }
+        DestroyEntity(self);
+    }
+}
+
+// falling rock with puff of smoke when it disappears. I think part of the
+// merman room breakable rock
+void EntityFallingRock(Entity* self) {
+    Collider collider;
+    Entity* newEntity;
+    s32 rnd;
+    s32 x, y;
+    s32 animFrame = self->params & 0xF;
+    s16 rndAngle;
+
+    switch (self->step) {
+    case 0:
+        InitializeEntity(D_us_80180A34);
+        self->animCurFrame = animFrame + 31;
+        self->drawFlags |= ENTITY_ROTATE | ENTITY_SCALEY | ENTITY_SCALEX;
+        self->scaleX = self->scaleY = 0x60;
+        rnd = (Random() & 0x1F) + 16;
+        rndAngle = (Random() * 6) + 0x900;
+        self->velocityX = rnd * rcos(rndAngle);
+        self->velocityY = rnd * rsin(rndAngle);
+        if (self->velocityX > 0) {
+            self->facingLeft = 1;
+        }
+        break;
+
+    case 1:
+        MoveEntity();
+        self->velocityY += FIX(0.125);
+        self->rotate -= 0x20;
+        x = self->posX.i.hi;
+        y = self->posY.i.hi + 8;
+        g_api.CheckCollision(x, y, &collider, 0);
+        if (collider.effects & EFFECT_SOLID) {
+            newEntity = AllocEntity(&g_Entities[224], &g_Entities[256]);
+            if (newEntity != NULL) {
+                CreateEntityFromEntity(E_INTENSE_EXPLOSION, self, newEntity);
+                newEntity->params = 0x10;
+                if (animFrame == 0) {
+                    newEntity->params = 0x13;
+                }
+            }
+            DestroyEntity(self);
+        }
+        break;
+    }
+}
+
